Rework frame colors (#23)
Introduce predictable, semantic colors.
Previously, colors were determined via the hash of frame names. While this is reproducible, it is not obvious to the user what this means.

To keep this viewer simple, we opt to select one color for Exec Frames and one color for all other frames.
We may introduce other colors and new meanings in the future.